// react-frontend/src/api/models/compact-user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CompactUserDTO
 */
export interface CompactUserDTO {
    /**
     * 
     * @type {number}
     * @memberof CompactUserDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof CompactUserDTO
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof CompactUserDTO
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof CompactUserDTO
     */
    'lastName'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof CompactUserDTO
     */
    'admin'?: boolean;
}